Detail 페이지 경로 잘못 입력 시 없는 데이터일 때 처리
